+++ crates/rust-client/src/sync/block_header.rs
@@ -43,7 +43,7 @@ impl<R: FeltRng> Client<R> {
 
     /// Attempts to retrieve the genesis block from the store. If not found,
     /// it requests it from the node and store it.
-    pub(crate) async fn ensure_genesis_in_place(&mut self) -> Result<BlockHeader, ClientError> {
+    pub async fn ensure_genesis_in_place(&mut self) -> Result<BlockHeader, ClientError> {
         let genesis = self.store.get_block_header_by_num(0.into()).await?;
 
         match genesis {
@@ -64,7 +64,7 @@ impl<R: FeltRng> Client<R> {
             MmrPeaks::new(0, vec![]).expect("Blank MmrPeaks should not fail to instantiate");
         // We specify that we want to store the MMR data from the genesis block as we might use it
         // as an anchor for created accounts.
-        self.store.insert_block_header(genesis_block, blank_mmr_peaks, true).await?;
+        self.store.insert_block_header(&genesis_block, blank_mmr_peaks, true).await?;
         Ok(genesis_block)
     }
 
@@ -147,7 +147,7 @@ impl<R: FeltRng> Client<R> {
 
         // Insert header and MMR nodes
         self.store
-            .insert_block_header(block_header, current_partial_mmr.peaks(), true)
+            .insert_block_header(&block_header, current_partial_mmr.peaks(), true)
             .await?;
         self.store.insert_chain_mmr_nodes(&path_nodes).await?;