+++ crates/web-client/src/models/merkle_path.rs
@@ -17,6 +17,7 @@ impl MerklePath {
         self.0.nodes().iter().map(Into::into).collect()
     }
 
+    #[wasm_bindgen(js_name = "computeRoot")]
     pub fn compute_root(&self, index: u64, node: &RpoDigest) -> RpoDigest {
         self.0.compute_root(index, node.clone().into()).unwrap().into()
     }