+++ crates/web-client/test/mocha.global.setup.mjs
@@ -5,6 +5,7 @@ import { spawn } from "child_process";
 
 import { register } from "ts-node";
 import { env } from "process";
+import { clearStore } from "./webClientTestUtils.js";
 
 chai.use(chaiAsPromised);
 
@@ -49,7 +50,7 @@ before(async () => {
 
   // Creates the client in the test context and attach to window object
   await testingPage.evaluate(
-    async (rpc_port, remote_prover_port) => {
+    async (rpcPort, remoteProverPort) => {
       const {
         Account,
         AccountHeader,
@@ -89,13 +90,13 @@ before(async () => {
         Word,
         WebClient,
       } = await import("./index.js");
-      let rpc_url = `http://localhost:${rpc_port}`;
-      let prover_url = null;
-      if (remote_prover_port) {
-        prover_url = `http://localhost:${remote_prover_port}`;
+      let rpcUrl = `http://localhost:${rpcPort}`;
+      let proverUrl = null;
+      if (remoteProverPort) {
+        proverUrl = `http://localhost:${remoteProverPort}`;
       }
       const client = new WebClient();
-      await client.create_client(rpc_url, prover_url);
+      await client.createClient(rpcUrl, proverUrl);
 
       window.client = client;
       window.Account = Account;
@@ -140,7 +141,7 @@ before(async () => {
       window.helpers = window.helpers || {};
 
       // Add the remote prover url to window
-      window.remote_prover_url = prover_url;
+      window.remoteProverUrl = proverUrl;
 
       window.helpers.waitForTransaction = async (
         transactionId,
@@ -153,12 +154,12 @@ before(async () => {
           if (timeWaited >= maxWaitTime) {
             throw new Error("Timeout waiting for transaction");
           }
-          await client.sync_state();
-          const uncomittedTransactions = await client.get_transactions(
+          await client.syncState();
+          const uncomittedTransactions = await client.getTransactions(
             window.TransactionFilter.uncomitted()
           );
           let uncomittedTransactionIds = uncomittedTransactions.map(
-            (transaction) => transaction.id().to_hex()
+            (transaction) => transaction.id().toHex()
           );
           if (!uncomittedTransactionIds.includes(transactionId)) {
             break;
@@ -170,7 +171,7 @@ before(async () => {
 
       window.helpers.refreshClient = async (initSeed) => {
         const client = new WebClient();
-        await client.create_client(rpc_url, prover_url, initSeed);
+        await client.createClient(rpcUrl, proverUrl, initSeed);
         window.client = client;
       };
     },
@@ -180,14 +181,7 @@ before(async () => {
 });
 
 beforeEach(async () => {
-  await testingPage.evaluate(async () => {
-    // Open a connection to the list of databases
-    const databases = await indexedDB.databases();
-    for (const db of databases) {
-      // Delete each database by name
-      indexedDB.deleteDatabase(db.name);
-    }
-  });
+  await clearStore();
 });
 
 after(async () => {