+++ CHANGELOG.md
@@ -5,15 +5,22 @@
 ### Features
 
 * Added support to import public accounts to `Client` (#733).
+* [BREAKING] Merged `TonicRpcClient` with `WebTonicRpcClient` and added missing endpoints (#744).
 
 ### Changes
 
+* Added wallet generation from seed & import from seed on web SDK (#710)
 * Add check for empty pay to id notes (#714).
 * [BREAKING] Refactored authentication out of the `Client` and added new separate authenticators (#718).
+* Added import/export for web client db (#740).
 * Re-exported RemoteTransactionProver in `rust-client` (#752).
 * Moved error handling to the `TransactionRequestBuilder::build()` (#750).
 * [BREAKING] Added starting block number parameter to `CheckNullifiersByPrefix` and removed nullifiers from `SyncState` (#758).
 * Added `ClientBuilder` for client initialization (#741).
+
+### Fixes
+
+* [BREAKING] Changed Snake Case Variables to Camel Case in JS/TS Files (#767).
 
 ## 0.7.0 (2025-01-28)